+++ libs/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: ibeliaie <ibeliaie@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/18 17:24:27 by ibeliaie          #+#    #+#             */
-/*   Updated: 2024/01/12 16:01:15 by ibeliaie         ###   ########.fr       */
+/*   Updated: 2024/02/11 18:28:25 by ibeliaie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,7 +53,7 @@ char	*ft_strnstr(const char *big, const char *smol, size_t len);
 
 int		ft_strncmp(const char *s1, const char *s2, size_t n);
 
-size_t	ft_strlen(const char *str);
+int		ft_strlen(const char *str);
 size_t	ft_strlcat(char *dst, const char *src, size_t dstsize);
 size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize);
